feature: add sprites rotation functionality
